rgw/admin/ratelimit_test.go
package admin

import (
	"context"
	"net/http"
	"testing"

	"github.com/stretchr/testify/assert"
)

func (suite *RadosGWTestSuite) TestRateLimit() {
	suite.SetupConnection()
	co, err := New(suite.endpoint, suite.accessKey, suite.secretKey, newDebugHTTPClient(http.DefaultClient))
	assert.NoError(suite.T(), err)

	suite.T().Run("set rate limit to user but user ID is empty", func(_ *testing.T) {
		err := co.SetUserRateLimit(context.Background(), RateLimitSpec{})
		assert.Error(suite.T(), err)
		assert.EqualError(suite.T(), err, errMissingUserID.Error())
	})

	suite.T().Run("get user rate limit but no user is specified", func(_ *testing.T) {
		_, err := co.GetUserRateLimit(context.Background(), RateLimitSpec{})
		assert.Error(suite.T(), err)
		assert.EqualError(suite.T(), err, errMissingUserID.Error())
	})
}
